1. servlet request dispatcher example 2. restructuring and package renaming
