+++ _02servlet-demo/servlet-demo/src/main/java/com/example/servletdemo/HelloServlet.java
@@ -1,4 +1,4 @@
-package com.example.servlet_demo;
+package com.example.servletdemo;
 
 import java.io.*;
 import javax.servlet.http.*;
